requirements.txt
accelerate==0.33.0
aiobotocore==2.16.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aioitertools==0.12.0
aiosignal==1.3.2
anndata==0.11.1
annotated-types==0.7.0
anyio==4.7.0
array-api-compat==1.9.1
arxiv==2.1.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
backports-tarfile==1.2.0
beautifulsoup4==4.12.3
biopython==1.84
biothings-client==0.3.1
botocore==1.35.81
cattrs==24.1.2
cellxgene-census==1.15.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
chembl-webresource-client==0.10.9
click==8.1.7
comm==0.2.2
contourpy==1.3.1
coredis==4.18.0
cryptography==44.0.0
cycler==0.12.1
dataclasses==0.8
datasets==2.20.0
decorator==5.1.1
deprecated==1.2.15
dill==0.3.8
distro==1.9.0
docstring-parser==0.16
docutils==0.21.2
easydict==1.13
et-xmlfile==2.0.0
evaluate==0.4.2
executing==2.1.0
feedparser==6.0.11
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.5.0
fuzzywuzzy==0.18.0
gget==0.29.0
h11==0.14.0
h5py==3.12.1
html2text==2024.2.26
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
importlib-metadata==8.5.0
ipython==8.30.0
ipywidgets==8.1.5
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.1.0
jedi==0.19.2
jeepney==0.8.0
jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab-widgets==3.0.13
keyring==25.5.0
kiwisolver==1.4.7
latexcodec==3.0.0
legacy-api-wrap==1.4.1
levenshtein==0.26.1
limits==3.14.1
litellm==1.53.3
llvmlite==0.43.0
lxml==5.3.0
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
moleculeace==3.0.0
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mygene==3.2.2
mysql-connector-python==9.1.0
natsort==8.4.0
networkx==3.4.2
nh3==0.2.20
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.57.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.0.0
pkginfo==1.12.0
platformdirs==4.3.6
prompt-toolkit==3.0.48
propcache==0.2.1
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyaml==24.12.1
pyarrow==18.1.0
pyarrow-hotfix==0.6
pybtex==0.24.0
pycparser==2.22
pydantic==2.10.3
pydantic-core==2.27.1
pydantic-settings==2.7.0
pydot==3.0.3
pygments==2.18.0
pympler==1.1
pymupdf==1.25.1
pynndescent==0.5.13
pyparsing==3.2.0
pytdc==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-levenshtein==0.26.1
pytz==2024.2
pyyaml==6.0.2
rapidfuzz==3.11.0
rdkit==2023.9.6
rdkit-pypi==2022.9.5
readme-renderer==44.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-cache==1.2.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rpds-py==0.22.3
s3fs==2024.5.0
safetensors==0.4.5
scanpy==1.10.4
scikit-learn==1.2.2
scikit-optimize==0.10.2
scipy==1.14.1
seaborn==0.13.2
secretstorage==3.3.3
session-info==1.0.0
setuptools==75.6.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
somacore==1.0.11
soupsieve==2.6
stack-data==0.6.3
statsmodels==0.14.4
stdlib-list==0.11.0
sympy==1.13.1
tantivy==0.22.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tiledb==0.29.1
tiledbsoma==1.11.4
tokenizers==0.19.1
torch==2.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.43.4
triton==3.1.0
twine==6.0.1
typing-extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
url-normalize==1.4.3
urllib3==2.2.3
wcwidth==0.2.13
widgetsnbextension==4.0.13
wrapt==1.17.0
xxhash==3.5.0
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
